+++ ChangeLog.md
@@ -20,6 +20,7 @@ Change type of `CERES` and `Value` as parametric to allow to change type of Str,
 ## 0.13.0.0 -- 2020-04-09
 
 ### Added
+* Add new `VariablePlace` `AtTricky` for special variable such as `WorldTime`, `siID`, etc.
 * Add new Array type `Value`
 
 ### Changed